added calendar and update panels
