+++ .config/ags/main.ts
@@ -1,5 +1,4 @@
-import NotificationPopups from "widgets/rightPanel/NotificationPopups";
-import NotificationPopups_NEW from "widgets/rightPanel/NotificationPopups_not_mine_and_too_complicated_for_no_reason";
+import NotificationPopups from "widgets/NotificationPopups";
 import RightPanel from "widgets/rightPanel/RightPanel";
 import Bar from "widgets/bar/Bar";
 import AppLauncher from "widgets/AppLauncher";
