+++ .config/ags/scss/lib.scss
@@ -102,7 +102,7 @@ separator {
   // padding: 0px 10px;
   margin: 5px 5px;
   background: transparent;
-  border-top: 2px dashed $module-foreground;
+  border-top: 1px dashed $module-foreground;
 }
 
 menu {
