+++ .config/ags/scss/right_panel/right-panel.scss
@@ -22,12 +22,12 @@
     }
   }
 
-  .notification-panel {
-    @extend .module;
+  .notification-history {
+    // @extend .module;
     margin: 0px 10px;
-    padding: 5px;
 
     .filter {
+      margin-bottom: 5px;
       .spotify {
         background: $green;
       }
@@ -52,37 +52,33 @@
         }
       }
     }
+    .notification {
+      padding: 10px;
+      min-width: 0px;
 
-    .clear {
-      @extend .danger;
-    }
+      .summary {
+        color: $foreground;
+        font-size: 1.2rem;
+      }
 
-    .notification-event {
-      .notification {
-        margin: 5px 0px;
-        padding: 10px;
-        min-width: 0px;
+      .icon {
+        min-width: 20px;
+        min-height: 20px;
 
-        .summary {
-          color: $foreground;
-          font-size: 1.2rem;
+        image {
+          font-size: 25px;
         }
 
-        .icon {
-          min-width: 20px;
-          min-height: 20px;
-
-          image {
-            font-size: 25px;
-          }
-
-          box {
-            min-width: 40px;
-            min-height: 40px;
-            border-radius: 10px;
-          }
+        box {
+          min-width: 40px;
+          min-height: 40px;
+          border-radius: 10px;
         }
       }
+    }
+    .clear {
+      @extend .danger;
+      // margin-top: 5p;
     }
   }
 }
