+++ .config/ags/scss/style.scss
@@ -15,6 +15,8 @@
 @import "widgets/media.scss";
 @import "widgets/progress.scss";
 @import "widgets/cava.scss";
+@import "widgets/calendar.scss";
+@import "widgets/update.scss";
 
 @import "wallpaper/style.scss";
