+++ .config/ags/scss/widgets/update.scss
@@ -1,0 +1,5 @@
+.update {
+  @extend .module;
+  margin: 0px 10px;
+  padding: 10px;
+}
