+++ .config/ags/widgets/Calendar.ts
@@ -1,0 +1,19 @@
+export default () => Widget.Box({
+    class_name: "calendar",
+    hexpand: true,
+    child: Widget.Calendar({
+        hexpand: true,
+        showDayNames: true,
+        showDetails: true,
+        showHeading: true,
+        showWeekNumbers: true,
+        detail: (self, y, m, d) =>
+        {
+            return `<span color="white">${d}</span>`
+        },
+        onDaySelected: ({ date: [y, m, d] }) =>
+        {
+            print(`${y}. ${m}. ${d}.`)
+        },
+    })
+})
