+++ .config/ags/widgets/rightPanel/RightPanel.ts
@@ -1,20 +1,23 @@
 
-import { Notification_ } from "./components/notification";
 import { Resources } from "widgets/Resources";
 import waifu, { WaifuVisibility } from "./components/waifu";
 import { globalMargin, rightPanelExclusivity, rightPanelVisibility, rightPanelWidth, waifuVisibility } from "variables";
 import { setOption } from "utils/options";
+import Calendar from "widgets/Calendar";
+import Update from "widgets/Update";
+import NotificationHistory from "./NotificationHistory";
 
-const Notifications = await Service.import("notifications")
+
 
 const maxRightPanelWidth = 600;
-const minRightPanelWidth = 300;
+const minRightPanelWidth = 200;
 
 function WindowActions()
 {
     return Widget.Box({
         class_name: "window-actions",
-        hpack: "end", spacing: 5
+        hpack: "end",
+        spacing: 5
     }, Widget.Button({
         label: "",
         class_name: "expand-window",
@@ -43,126 +46,26 @@ function WindowActions()
 }
 
 
-interface Filter
-{
-    name: string
-    class: string
-}
-
-const notificationFilter = Variable<Filter>({ name: "", class: "" });
-
-function Filter()
-{
-    const Filters: Filter[] = [{
-        name: "Spotify",
-        class: "spotify",
-    }, {
-        name: "Clipboard",
-        class: "clipboard",
-    }, {
-        name: "Update",
-        class: "update",
-    }];
-
-    return Widget.Box({
-        class_name: "filter",
-        hexpand: false,
-        children: Filters.map(filter =>
-        {
-            return Widget.Button({
-                label: filter.name,
-                hexpand: true,
-                on_clicked: () => notificationFilter.value = (notificationFilter.value === filter ? { name: "", class: "" } : filter),
-                class_name: notificationFilter.bind().as(filter => filter.class),
-            })
-        })
-    })
-}
-
-
-
-
-
-function FilterNotifications(notifications: any[], filter: string): any[]
-{
-    const filtered: any[] = [];
-    const others: any[] = [];
-
-    notifications.forEach((notification: any) =>
-    {
-        if (notification.app_name.includes(filter) || notification.summary.includes(filter)) {
-            filtered.unshift(notification);
-        } else {
-            others.unshift(notification);
-        }
-    });
-    return [...filtered, ...others].slice(0, 50); // Limit to the last 50 notifications DEFAULT, higher number will slow down the UI
-}
-
-const NotificationHistory = () => Widget.Box({
-    vertical: true,
-    children: Utils.merge([notificationFilter.bind(), Notifications.bind("notifications")], (filter, notifications) =>
-    {
-        if (!notifications) return [];
-        return FilterNotifications(notifications, filter.name)
-            .map(notification =>
-            {
-                return Widget.EventBox(
-                    {
-                        class_name: "notification-event",
-                        on_primary_click: () => Utils.execAsync(`wl-copy "${notification.body}"`).catch(err => print(err)),
-                        child: Notification_(notification),
-                    });
-            })
-    }),
-})
-
-const Separator = () => Widget.Separator({ vertical: false });
-
-const NotificationsDisplay = Widget.Scrollable({
-    class_name: "notification-history",
-    hscroll: 'never',
-    vexpand: true,
-    child: NotificationHistory(),
-})
-
-const NotificationPanel = () =>
-{
-    return Widget.Box({
-        class_name: "notification-panel",
-        // spacing: 5,
-        vertical: true,
-        children: [Filter(), NotificationsDisplay, ClearNotifications()],
-    })
-}
-
-
-const ClearNotifications = () =>
-{
-    return Widget.Button({
-        class_name: "clear",
-        label: "Clear",
-        on_clicked: () =>
-        {
-            NotificationsDisplay.child = Widget.Box({
-                vertical: true,
-                children: [],
-            });
-            Notifications.clear()
-            // .finally(() => NotificationsDisplay.child = NotificationHistory())
-        },
-    })
-}
-
 function Panel()
 {
     return Widget.Box({
         css: rightPanelWidth.bind().as(width => `*{min-width: ${width}px}`),
         vertical: true,
         // spacing: 5,
-        children: [WindowActions(), waifu(), Separator(), Resources(), Separator(), NotificationPanel()],
+        children: [
+            WindowActions(),
+            waifu(),
+            Separator(),
+            Resources(),
+            Separator(),
+            Update(),
+            Separator(),
+            NotificationHistory()
+        ],
     })
 }
+
+const Separator = () => Widget.Separator({ vertical: false });
 
 const Window = () => Widget.Window({
     name: `right-panel`,
